increase precision on mobile

+++ sketch/landscape-with-primitive-shapes/assets/primitive-shapes-landscape.frag
@@ -1,8 +1,5 @@
-// Author: Claude
-// Title: Perspective Dots
-
 #ifdef GL_ES
-precision mediump float;
+precision highp float;
 #endif
 
 uniform vec2 u_resolution;
